Offer create form moved to offers directory
